requirements.txt
pip~=25.1
emoji~=2.14.1
attrs~=24.3.0
zlib~=1.2.13
distro~=1.9.0
wheel~=0.45.1
pysocks~=1.7.1
protobuf~=5.29.5
openssl~=3.0.16
cryptography~=45.0.3
pillow~=11.3.0
tornado~=6.5.1
jinja2~=3.1.6
filelock~=3.18.0
setuptools~=78.1.1
packaging~=24.2
pyparsing~=3.2.3
build~=1.2.2.post1
matplotlib~=3.10.3
streamlit~=1.46.1
python-dotenv~=1.1.1
langchain-core~=0.3.68
langchain-community~=0.3.27
langchain~=0.3.26
sympy~=1.14.0